Rename: auth.ts -> authReducer.ts
